move extracting images from terraform to pkg extractimages
